// tims-frontend/src/app/shared/models/customer.ts
export class Customer {
    id: number;
    salutation: string;
    firstname: string;
    lastname: string;
    contactno: string;
    alternatecontact: string;
    emailid: string;
    alternateemail: string;
    faxno: string;
    billstreetno: string;
    billstreetname: string;
    billcity: string;
    billpostalcode: string;
    billstate: string;
    billcountry: string;
    shipstreetno: string;
    shipstreetname: string;
    shipcity: string;
    shippostalcode: string;
    shipstate: string;
    shipcountry: string;

    constructor(){
        
    }
}